+++ SRC/Waifu2x-Extension-QT/waifu2x_caffe.cpp
@@ -25,7 +25,7 @@ int MainWindow::Waifu2x_Caffe_Image(int rowNum)
     //============================= 读取设置 ================================
     int ScaleRatio = ui->spinBox_ScaleRatio_image->value();
     int DenoiseLevel = ui->spinBox_DenoiseLevel_image->value();
-    bool DelOriginal = ui->checkBox_DelOriginal->isChecked();
+    bool DelOriginal = (ui->checkBox_DelOriginal->isChecked()||ui->checkBox_ReplaceOriginalFile->isChecked());
     bool SaveAsJPG = ui->checkBox_SaveAsJPG->isChecked();
     //======
     if(SaveAsJPG)
@@ -263,13 +263,16 @@ int MainWindow::Waifu2x_Caffe_Image(int rowNum)
     }
     if(DelOriginal)
     {
-        if(ui->checkBox_Move2RecycleBin->isChecked())
+        if(ReplaceOriginalFile(SourceFile_fullPath,OutPutPath_Final)==false)
         {
-            file_MoveToTrash(SourceFile_fullPath);
-        }
-        else
-        {
-            QFile::remove(SourceFile_fullPath);
+            if(ui->checkBox_Move2RecycleBin->isChecked())
+            {
+                file_MoveToTrash(SourceFile_fullPath);
+            }
+            else
+            {
+                QFile::remove(SourceFile_fullPath);
+            }
         }
         status = "Finished, original file deleted";
         emit Send_Table_image_ChangeStatus_rowNumInt_statusQString(rowNum, status);
@@ -301,7 +304,7 @@ int MainWindow::Waifu2x_Caffe_GIF(int rowNum)
     //============================= 读取设置 ================================
     int ScaleRatio = ui->spinBox_ScaleRatio_gif->value();
     int DenoiseLevel = ui->spinBox_DenoiseLevel_gif->value();
-    bool DelOriginal = ui->checkBox_DelOriginal->isChecked();
+    bool DelOriginal = (ui->checkBox_DelOriginal->isChecked()||ui->checkBox_ReplaceOriginalFile->isChecked());
     bool OptGIF = ui->checkBox_OptGIF->isChecked();
     int Sub_gif_ThreadNumRunning = 0;
     QString OutPutPath_Final ="";
@@ -503,13 +506,16 @@ int MainWindow::Waifu2x_Caffe_GIF(int rowNum)
     //============================= 删除原文件 &  & 更新table status ============================
     if(DelOriginal)
     {
-        if(ui->checkBox_Move2RecycleBin->isChecked())
+        if(ReplaceOriginalFile(SourceFile_fullPath,OutPutPath_Final)==false)
         {
-            file_MoveToTrash(SourceFile_fullPath);
-        }
-        else
-        {
-            QFile::remove(SourceFile_fullPath);
+            if(ui->checkBox_Move2RecycleBin->isChecked())
+            {
+                file_MoveToTrash(SourceFile_fullPath);
+            }
+            else
+            {
+                QFile::remove(SourceFile_fullPath);
+            }
         }
         status = "Finished, original file deleted";
         emit Send_Table_gif_ChangeStatus_rowNumInt_statusQString(rowNum, status);
@@ -629,7 +635,7 @@ int MainWindow::Waifu2x_Caffe_Video(int rowNum)
     //============================= 读取设置 ================================
     int ScaleRatio = ui->spinBox_ScaleRatio_video->value();
     int DenoiseLevel = ui->spinBox_DenoiseLevel_video->value();
-    bool DelOriginal = ui->checkBox_DelOriginal->isChecked();
+    bool DelOriginal = (ui->checkBox_DelOriginal->isChecked()||ui->checkBox_ReplaceOriginalFile->isChecked());
     bool isCacheExists = false;
     bool isVideoConfigChanged = true;
     int Sub_video_ThreadNumRunning = 0;
@@ -952,13 +958,16 @@ int MainWindow::Waifu2x_Caffe_Video(int rowNum)
     //============================= 删除原文件 &  & 更新table status ============================
     if(DelOriginal)
     {
-        if(ui->checkBox_Move2RecycleBin->isChecked())
+        if(ReplaceOriginalFile(SourceFile_fullPath,OutPutPath_Final)==false)
         {
-            file_MoveToTrash(SourceFile_fullPath);
-        }
-        else
-        {
-            QFile::remove(SourceFile_fullPath);
+            if(ui->checkBox_Move2RecycleBin->isChecked())
+            {
+                file_MoveToTrash(SourceFile_fullPath);
+            }
+            else
+            {
+                QFile::remove(SourceFile_fullPath);
+            }
         }
         status = "Finished, original file deleted";
         emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
@@ -994,7 +1003,7 @@ int MainWindow::Waifu2x_Caffe_Video_BySegment(int rowNum)
     //============================= 读取设置 ================================
     int ScaleRatio = ui->spinBox_ScaleRatio_video->value();
     int DenoiseLevel = ui->spinBox_DenoiseLevel_video->value();
-    bool DelOriginal = ui->checkBox_DelOriginal->isChecked();
+    bool DelOriginal = (ui->checkBox_DelOriginal->isChecked()||ui->checkBox_ReplaceOriginalFile->isChecked());
     bool isCacheExists = false;
     bool isVideoConfigChanged = true;
     int Sub_video_ThreadNumRunning = 0;
@@ -1480,13 +1489,16 @@ int MainWindow::Waifu2x_Caffe_Video_BySegment(int rowNum)
     //============================= 删除原文件 &  & 更新table status ============================
     if(DelOriginal)
     {
-        if(ui->checkBox_Move2RecycleBin->isChecked())
+        if(ReplaceOriginalFile(SourceFile_fullPath,OutPutPath_Final)==false)
         {
-            file_MoveToTrash(SourceFile_fullPath);
-        }
-        else
-        {
-            QFile::remove(SourceFile_fullPath);
+            if(ui->checkBox_Move2RecycleBin->isChecked())
+            {
+                file_MoveToTrash(SourceFile_fullPath);
+            }
+            else
+            {
+                QFile::remove(SourceFile_fullPath);
+            }
         }
         status = "Finished, original file deleted";
         emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
